Add Core Unit tabs items
[X] Show the correct items for tab core-units

What solved
